feat: logger , fix: download big files

+++ src/config/Page.cpp
@@ -55,7 +55,7 @@ Page *Page::fromNode(Node<Token *> *root) {
     {
         throw IllegalArgumentException("root must be a code node");
     }
-    for (int i = 0; i< root->getChildren().size(); i++)
+    for (size_t i = 0; i< root->getChildren().size(); i++)
     {
 
                if (is_digits(root->getChildren()[i]->getData()->getValue()))
